requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
pillow==11.1.0
sqlparse==0.5.3
django-admin-sortable2==2.1.7
django-cors-headers==4.3.1
python-dotenv==1.0.1